requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
gradio>=4.44.1
requests>=2.31.0
python-dotenv>=0.19.0 